Move `data_channel_open::Error` to `error::DataChannelOpenError`
